Updated to be release neutral
 - Added a mandatory -ReleaseNum param
 - Added mechanism to figure determine filename to download (different per release)
 - Added `dnf update` before installing packages to ensure latest version compatibility
 - Added `dnf-plugins-core` to package install which provides `dnf copr` command
 - Commented out parts dealing with removing Fedora trademarked items (note below)
 - Removed password for new user, which is consistent with Fedora containers and Fedora Toolbox

Trademarks:  In Fedora 36, your lines of code were problematic since there were changes to these packages.  Both `fedora-logos` and `fedora-release-notes` were removed, and I suspect were integrated with `fedora-release`.  But this was workable since I could have them skipped if not present, however in Fedora 36, the `fedora-release` was tied to `sudo` for some reason and I didn't feel it was worth the risk of forcing the removal, at least across multiple versions, so I just removed the  parts of the code that was trying to switch to the generic packages.

On a side note, I really don't think you need to worry about removing the trademarked packages since you are only providing an installer.   If you wanted to host Fedora on the Microsoft Store, then I think it would be necessary to make that change, but not with what you are doing.  And as a very small sample of your userbase, I personally would not want those packages removed from my install in WSL.  I think it best to leave them and let users manually remove the trademarked content after the install.

BTW, thanks for you efforts in writing this script!!!  It saved me from having to write it myself.  :-)  Hope the additions help.  I verified it worked on Fedora 33, 34, 35, and 36.  I would have went back further, but Wslu repo only went back to Fedora 33.